Allow fractional degrees (and hours) and minutes in String#dms_to_d (and String#hms_to_d).  What could go wrong?  :-)
git-svn-id: https://svn.hcro.org/mmm@746 9b194dde-2b4d-0410-98f0-bba1668eb47e

+++ davidm/astroutil/lib/astroutil.rb
@@ -111,7 +111,7 @@ class String
   # <b>NOT</b> the inverse of Numeric#to_dmsstr (but may become so).
   def dms_to_d;
     d, m, s = split(':',3)
-    d, m, s = d.to_i, m.to_i, s.to_f
+    d, m, s = d.to_f, m.to_f, s.to_f
     m, s = -m, -s if d == 0 && self =~ /^\s*-/
     Math::dms_to_d(d,m,s)
   end
